+++ XF/Pub/Controller/Conversation.php
@@ -30,6 +30,10 @@ class Conversation extends XFCP_Conversation
 
         $breadcrumbs = [
             [
+                'value' => \XF::phrase('conversations'),
+                'href' => $this->buildLink('conversations'),
+            ],
+            [
                 'value' => $message->Conversation->title,
                 'href' => $this->buildLink(
                     'conversations',
@@ -49,8 +53,8 @@ class Conversation extends XFCP_Conversation
     }
 
     /**
-     * @param int $messageId
-     * @param array $extraWith
+     * @param int      $messageId
+     * @param string[] $extraWith
      *
      * @return \XF\Entity\ConversationMessage
      *