+++ _output/extension_hint.php
@@ -26,6 +26,11 @@ namespace Jrahmy\Moderation\XF\Pub\Controller
 	class XFCP_Report extends \XF\Pub\Controller\Report {}
 }
 
+namespace Jrahmy\Moderation\XF\Report
+{
+	class XFCP_ConversationMessage extends \XF\Report\ConversationMessage {}
+}
+
 namespace Jrahmy\Moderation\XF\Repository
 {
 	class XFCP_UserAlert extends \XF\Repository\UserAlert {}
@@ -34,6 +39,7 @@ namespace Jrahmy\Moderation\XF\Repository
 namespace Jrahmy\Moderation\XF\Service\Report
 {
 	class XFCP_Commenter extends \XF\Service\Report\Commenter {}
+	class XFCP_CommentPreparer extends \XF\Service\Report\CommentPreparer {}
 	class XFCP_Creator extends \XF\Service\Report\Creator {}
 	class XFCP_Notifier extends \XF\Service\Report\Notifier {}
 }