+++ CHANGELOG.md
@@ -4,6 +4,34 @@ Changelog
 Unreleased
 ----------
 
+[1.0.3] - 2020-01-06
+--------------------
+### Added
+- Display content link on conversation message reports
+- Implement editing for report comments
+- Implement deletion for report comments
+- Implement IP logging for report comments
+- Implement reactions for report comments
+
+### Changed
+- Fix conversation warning breadcrumbs
+- Code clean-up
+
+[1.0.2] - 2019-12-21
+--------------------
+### Added
+- Link reports and warnings together
+- Implement conversation message warnings
+
+[1.0.1] - 2019-12-16
+--------------------
+### Added
+- Redirect to a report after it is created if the visitor can view it
+- Add a 'View report' button to content with an existing report
+
+### Changed
+- Code clean-up
+
 [1.0.0] - 2019-11-21
 --------------------
 ### Changed